+++ src/Core/Domain/Entities/Book.cs
@@ -13,7 +13,7 @@ public class Book : Entity
     public string? Title { get; init; }
     public string? Description { get; init; }
     public DateTime? PublishDate { get; init; }
-    public int PagesCount { get; init; }
+    public short PagesCount { get; init; }
     public string? Language { get; init; }
     public Guid AuthorId { get; init; }
 }
